[LED-113] Move out Export Code
